psp-8638 ensure parcel map layer displays immediately.
