+++ source/backend/api/Pims.Api.csproj
@@ -2,9 +2,9 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
     <UserSecretsId>0ef6255f-9ea0-49ec-8c65-c172304b4926</UserSecretsId>
-    <Version>5.3.1-81.35</Version>
-    <Version>5.3.1-81.35</Version>
-    <AssemblyVersion>5.3.1.81</AssemblyVersion>
+    <Version>5.3.2-81.35</Version>
+    <Version>5.3.2-81.35</Version>
+    <AssemblyVersion>5.3.2.81</AssemblyVersion>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <ProjectGuid>16BC0468-78F6-4C91-87DA-7403C919E646</ProjectGuid>
     <TargetFramework>net8.0</TargetFramework>
