+++ source/frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "frontend",
-  "version": "5.3.1-81.35",
+  "version": "5.3.2-81.35",
   "private": true,
   "dependencies": {
     "@bcgov/bc-sans": "1.0.1",
